Move useBasicWidgetState hook
