+++ frontend/lib/src/components/widgets/Selectbox/Selectbox.tsx
@@ -26,7 +26,7 @@ import {
 import {
   useBasicWidgetState,
   ValueWithSource,
-} from "@streamlit/lib/src/useBasicWidgetState"
+} from "@streamlit/lib/src/hooks/useBasicWidgetState"
 
 export interface Props {
   disabled: boolean
