+++ frontend/lib/src/hooks/useBasicWidgetState.ts
